Remove prepublish from npm
This gets run, inexplicably, during a call to npm i.

In combination with the jshint call in this module's prepublish, it
means that if you go to install the module, it'll run jshint, which
silently walks up the source tree until it finds a jshintrc. In our
case, this caused jshint to fail because the repository in question is
using a strict jshintrc.

This caused NPM to silently fail to install this module in to our
repository.